Update patch set 1
Patch Set 1:

This may not be a perfect solution, but it does prevent the case mentioned in the issue.  I chose not to explicitly check for the scheme because there are cases where it's valid to not have a scheme.  For example on the Gitlab server at my company, urls begin with 'git@'.

Patch-set: 1
Reviewer: David Pursehouse <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0