Update patch set 1
Patch Set 1:

Yes, this approach is much better. And preserving GIT_ALLOW_PROTOCOL if the user has it already set is nice.

Adding git-remote-persistent-https seems fine. Is git-remote-sso an internal Google protocol? But anyways, if any URL is safe to pass to those, they should be fine to add to the whitelist.

Patch-set: 1