Update patch set 1
Patch Set 1:

There's already two implementations of basically this functionality in the code already. They just trigger in different circumstances :)

The --optimized_fetch option to sync will skip the network sync (with need_to_fetch in project.py) if the revisionExpr is a SHA1.

If -c (current branch only) is set, and revisionExpr is a SHA1, we exit from _RemoteFetch early.

These should probably be updated to use revisionId (and make sure everything else mixes revisionExpr and revisionId properly)

Patch-set: 1