Update patch set 1
Patch Set 1:

> There's also _CheckForSha1, which has a similar purpose (but looks
 > directly at the revisionExpr)

It is different in 2 ways:

1) it looks at revisionExpr (not an arbitrary given), we could change that easily

2) it uses git rev-parse, which does a completely different thing. I used to think it would check for the sha1 being there, but no.

--verify
           Verify that exactly one parameter is provided, and that it can be turned into a raw 20-byte SHA-1 that can be used to access the object database. If so, emit it to the standard output;
           otherwise, error out.

           If you want to make sure that the output actually names an object in your object database and/or can be used as a specific type of object you require, you can add the ^{type} peeling
           operator to the parameter. For example, git rev-parse "$VAR^{commit}" will make sure $VAR names an existing object that is a commit-ish (i.e. a commit, or an annotated tag that points at a
           commit). To make sure that $VAR names an existing object of any type, git rev-parse "$VAR^{object}" can be used.

       -q, --quiet
           Only meaningful in --verify mode. Do not output an error message if the first argument is not a valid object name; instead exit with non-zero status silently. SHA-1s for valid object names
           are printed to stdout on success.

So in the way `git rev-parse --verify -q <sha1>` it only checks if the sha1 is "valid" syntax, i.e. not exceeding 40 chars, no non-hex digits.

It *doesn't* check for existence of the underlying object. To do that we'd need to do append a `^{commit}` IIUC. 

Looking at the use cases of _CheckForSha1 it seems to me as if it wants to do a similar thing (Check that the object exists that revisionExpr points to), but there might be a bug of checking the existence of the object vs checking the syntax?

Patch-set: 1