Update patch set 1
Patch Set 1:

Both of them are ways to say "if we already know what SHA1 this project should be at, and we have it, don't talk to the server". We don't turn it on by default, since people sometimes need to pull down other branches (or want newer git history).

The common case where the revisionExpr is a sha1 is during smartsync or build server builds, where you've got a generated manifest that hardcodes all of the SHA1s, with the branch names being in "upstream".

As long as we fix the way to understand that we've got local SHA1, -c, --optimized-fetch, and -Q would behave the same for repositories that have a local SHA1.

The difference between -c and --optimized-fetch is for repos that don't have the local SHA1, whether we download the other branches or not.

If we were to start over, for the common usecases, I'd strongly consider making -c the default, and have to explicitly request all branches/more history, just so we could make these optimizations the default.

Patch-set: 1
Reviewer: Dan Willemsen <1021072@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0