Update patch set 2
Patch Set 2:

c/83694 is not related. I cherry-picked that and tested it, but it only deals with the `git push` URL, which is not used at all by the `repo upload` command.

It might, however, relate to this: https://gerrit-review.googlesource.com/#/c/76662/3 - I haven't tested that yet, but based on the code I think it might be an alternative approach.

Patch-set: 2
Label: Verified=0