Update patch set 2
Patch Set 2:

(1 comment)

> Which version of git are you using? Are you able to fetch a non-tip shallow clone from this server?

We are using Gerrit 2.12.4 and if I try your command on our server and one of our projects it fails. This is working on your server and I suppose you're also using Gerrit.

So maybe there is some parameter that can be set in Gerrit server to make this works. Please tell me if it's the case.

Patch-set: 2