Update patch set 1
Patch Set 1: Code-Review+1

> It's still related and more useful than nothing at all. insteadOf
 > merely adds one (static and typically obvious) indirection layer
 > but the information is still made available.

I think Dan is asking if the insteadOf redirection is accounted for in this message. Is it?

I think this is still better than nothing even if it doesn't, but it's good to know the limitations.

Patch-set: 1
Reviewer: Nasser Grainawi <1002666@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1