Update patch set 1
Patch Set 1:

it seems weird that you want to log the URL that is failing to help people track down issues, but then you don't want to log the actual URL that is failing, but a diff one.  so if you're hitting this failure and you have insteadOf active, you'll get one URL displayed, and you'll bang on it and everything works, and then you'll curse repo for being stupid.

so if your claim is "repo's failing output atm is not helpful for further triaging", i'd agree, but i'm not sure this change improves things overall if insteadOf is active.  it points you in the wrong direction :).

Patch-set: 1
Reviewer: Mike Frysinger <1007145@173816e5-2b9a-37c3-8a2e-48639d4f1153>