Update patch set 2
Patch Set 2: Verified+1

> Patch Set 1:
> 
> project.py is an internal API that reflects the Gerrit interface.  subcmds/ is for user interfaces.  so project.py should have a notify setting, not a pseudo "silent" option.

modified project.py API to accept a notify string. Did not add a check for the string being a valid option, not sure if it is ok to rely on the server in this case.

Patch-set: 2
Reviewer: Gerrit User 1014182 <1014182@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1