Update patch set 1
Patch Set 1:

> the API of the funcs are simple: they take a % formatted string

That's exactly the point: With the minimum Python version required by repo being 2.7 by now, I was mildly surprised for repo to still use the legacy %-formatting syntax instead of format().

> in fact, the API is already quite clear in that regard by virtue of the fact that it takes *args.

Again, the pure fact that the API takes *args does not mean it's clear that is takes an %-formatted string. It could also have been an format()-string with {0}-style placeholders.

> if the caller is surprised by the fact that ('%s foo', 'hi') turns into 'hi foo', then they aren't actually reading the code.  if the caller isn't surprised by that, but is surprised that passing '%s foo' causes a crash, then they aren't paying attention.

Neither of these was my use-case. My use-case was that I was surprised that ('%20') crashes, whereas that's clearly not even a valid %-formatting string, so I simply would have expected no variable interpolation to happen. But after all, I'm not a Python guy.

> if the caller passed '%s' but forgot to pass a 2nd arg, your proposal would allow that mistake.  if the caller passed '%s ' + cwd as the 1st arg, your proposal wouldn't catch that mistake.

Right, with my change there is a small risk that if the caller forgets to pass a 2nd argument the result is not what he/she expects as the formatting string does not get expanded. But at least it does not crash like it does now on improper use.

Patch-set: 1
Reviewer: Gerrit User 1015513 <1015513@173816e5-2b9a-37c3-8a2e-48639d4f1153>