Update patch set 2
Patch Set 2:

I generally agree, the output for 'revision' now looks correct.

However, I have a CI flow that relies on the output of the manifest branch for a given project. I don't think there is a way to obtain it now, would you take a patch that re-adds it as a different field?

Patch-set: 2
CC: Gerrit User 1032632 <1032632@173816e5-2b9a-37c3-8a2e-48639d4f1153>